fixed some errors; added support template {ext} when saving images, added .php_cs config

+++ tests/SaveGoogleImagesTest.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Nelexa\GPlay\Tests;
@@ -8,42 +9,44 @@ use Nelexa\GPlay\GPlayApps;
 use Nelexa\GPlay\Model\GoogleImage;
 use Nelexa\GPlay\Model\ImageInfo;
 use PHPUnit\Framework\TestCase;
-use RecursiveDirectoryIterator;
-use RecursiveIteratorIterator;
-use Symfony\Component\Cache\Simple\ArrayCache;
 
-class SaveGoogleImagesTest extends TestCase
+/**
+ * @internal
+ *
+ * @small
+ */
+final class SaveGoogleImagesTest extends TestCase
 {
-    /**
-     * @var string
-     */
+    /** @var string */
     private $destDirectory;
-    /**
-     * @var GPlayApps
-     */
+
+    /** @var GPlayApps */
     private $gplay;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->destDirectory = sys_get_temp_dir() . '/gplay-screenshots';
         $this->gplay = new GPlayApps();
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         parent::tearDown();
 
-        $it = new RecursiveDirectoryIterator($this->destDirectory, RecursiveDirectoryIterator::SKIP_DOTS);
-        $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
-        foreach ($files as $file) {
-            if ($file->isDir()) {
-                rmdir($file->getRealPath());
-            } else {
-                unlink($file->getRealPath());
+        if (is_dir($this->destDirectory)) {
+            $it = new \RecursiveDirectoryIterator($this->destDirectory, \RecursiveDirectoryIterator::SKIP_DOTS);
+            $files = new \RecursiveIteratorIterator($it, \RecursiveIteratorIterator::CHILD_FIRST);
+
+            foreach ($files as $file) {
+                if ($file->isDir()) {
+                    rmdir($file->getRealPath());
+                } else {
+                    unlink($file->getRealPath());
+                }
             }
+            rmdir($this->destDirectory);
         }
-        rmdir($this->destDirectory);
     }
 
     /**
@@ -55,19 +58,18 @@ class SaveGoogleImagesTest extends TestCase
         $icon = $detailApp->getIcon();
 
         $icon->setSize(128);
-        $imageInfo = $icon->saveAs($this->destDirectory . '/icon_small.png');
-        $this->assertInstanceOf(ImageInfo::class, $imageInfo);
-        $this->assertContains(128, [$imageInfo->getWidth(), $imageInfo->getHeight()]);
-
+        $imageInfo = $icon->saveAs($this->destDirectory . '/icon_small.{ext}');
+        self::assertInstanceOf(ImageInfo::class, $imageInfo);
+        self::assertContains(128, [$imageInfo->getWidth(), $imageInfo->getHeight()]);
         $icon->setSize(256);
-        $imageInfo = $icon->saveAs($this->destDirectory . '/icon_medium.png');
-        $this->assertInstanceOf(ImageInfo::class, $imageInfo);
-        $this->assertContains(256, [$imageInfo->getWidth(), $imageInfo->getHeight()]);
+        $imageInfo = $icon->saveAs($this->destDirectory . '/icon_medium.{ext}');
+        self::assertInstanceOf(ImageInfo::class, $imageInfo);
+        self::assertContains(256, [$imageInfo->getWidth(), $imageInfo->getHeight()]);
 
         $icon->setSize(512);
-        $imageInfo = $icon->saveAs($this->destDirectory . '/icon_large.png');
-        $this->assertInstanceOf(ImageInfo::class, $imageInfo);
-        $this->assertContains(512, [$imageInfo->getWidth(), $imageInfo->getHeight()]);
+        $imageInfo = $icon->saveAs($this->destDirectory . '/icon_large.{ext}');
+        self::assertInstanceOf(ImageInfo::class, $imageInfo);
+        self::assertContains(512, [$imageInfo->getWidth(), $imageInfo->getHeight()]);
     }
 
     /**
@@ -77,9 +79,12 @@ class SaveGoogleImagesTest extends TestCase
     {
         $detailApp = $this->gplay->getApp('com.rovio.angrybirdsfriends');
         $screenshots = $detailApp->getScreenshots();
-        array_walk($screenshots, static function (GoogleImage $image) {
-            $image->setSize(512);
-        });
+        array_walk(
+            $screenshots,
+            static function (GoogleImage $image): void {
+                $image->setSize(512);
+            }
+        );
 
         $destDir = $this->destDirectory;
 
@@ -87,38 +92,50 @@ class SaveGoogleImagesTest extends TestCase
             $screenshots,
             static function (GoogleImage $image) use ($destDir) {
                 $hashUrl = $image->getHashUrl('md5', 2);
-                return $destDir . DIRECTORY_SEPARATOR . $hashUrl . '.png';
+
+                return $destDir . \DIRECTORY_SEPARATOR . $hashUrl . '.{ext}';
             }
         );
-        $this->assertIsArray($imageInfos);
-        $this->assertCount(count($screenshots), $imageInfos);
-        $this->assertContainsOnlyInstancesOf(ImageInfo::class, $imageInfos);
+        self::assertIsArray($imageInfos);
+        self::assertCount(\count($screenshots), $imageInfos);
+        self::assertContainsOnlyInstancesOf(ImageInfo::class, $imageInfos);
     }
 
     /**
      * @throws GooglePlayException
      */
-    public function testImages(): void
+    public function testSaveNotFoundImage(): void
     {
         $this->expectException(GooglePlayException::class);
         $this->expectExceptionMessage('404 Not Found');
 
-        $invalidImage = new GoogleImage('https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa000');
+        $invalidImage = new GoogleImage(
+            'https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa000'
+        );
         $invalidImage->saveAs($this->destDirectory . '/images.png');
     }
 
-    public function testSaveInvalidImages(): void
+    public function testSaveNotFoundImages(): void
     {
         $this->expectException(GooglePlayException::class);
 
-        $invalidImage = new GoogleImage('https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa000');
-        $invalidImage2 = new GoogleImage('https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa002');
-        $invalidImage3 = new GoogleImage('https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa003');
+        $invalidImage = new GoogleImage(
+            'https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa000'
+        );
+        $invalidImage2 = new GoogleImage(
+            'https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa002'
+        );
+        $invalidImage3 = new GoogleImage(
+            'https://lh3.googleusercontent.com/DKoidc0T3T1KvYC2stChcX9zwmjKj1pgmg3hXzGBDQXM8RG_7JjgiuS0CLOh8DUa003'
+        );
 
         $images = [$invalidImage, $invalidImage2, $invalidImage3];
-        array_walk($images, static function (GoogleImage $image) {
-            $image->setSize(512);
-        });
+        array_walk(
+            $images,
+            static function (GoogleImage $image): void {
+                $image->setSize(512);
+            }
+        );
 
         $destDir = $this->destDirectory;
 
@@ -126,7 +143,8 @@ class SaveGoogleImagesTest extends TestCase
             $images,
             static function (GoogleImage $image) use ($destDir) {
                 $hashUrl = $image->getHashUrl('crc32');
-                return $destDir . DIRECTORY_SEPARATOR . $hashUrl . '.png';
+
+                return $destDir . \DIRECTORY_SEPARATOR . $hashUrl . '.png';
             }
         );
     }
